+++ src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.WinForms/Form1.cs
@@ -4,24 +4,15 @@ namespace Sandbox.WinForms;
 
 public partial class Form1 : Form
 {
-    System.Windows.Forms.Timer Timer = new() { Interval = 10 };
-    ScottPlot.DataGenerators.RandomWalker2D Walker = new();
     public Form1()
     {
         InitializeComponent();
 
-        var logger = formsPlot1.Plot.Add.DataLogger();
-        logger.Add(0, 0);
-        logger.LineWidth = 2;
-
-        Timer.Tick += (s, e) =>
-        {
-            Coordinates pt = Walker.Next();
-            logger.Add(pt);
-            formsPlot1.Refresh();
-        };
-
-        Timer.Start();
-
+        var mrk = formsPlot1.Plot.Add.Marker(5, 5);
+        mrk.Size = 100;
+        mrk.Color = Colors.Black.WithAlpha(0.5f);
+        mrk.MarkerLineWidth = 2;
+        mrk.MarkerLineColor = Colors.Black;
+        mrk.Shape = MarkerShape.FilledCircle;
     }
 }