+++ src/ScottPlot5/ScottPlot5 Tests/Unit Tests/RenderTests/Plottable/MarkerTests.cs
@@ -30,16 +30,19 @@ internal class MarkerTests
             mp1.MarkerStyle.Shape = markerShapes[i];
             mp1.MarkerStyle.Size = 20;
             mp1.MarkerStyle.LineWidth = 2;
-            mp1.MarkerStyle.LineColor = Colors.Black;
-            mp1.MarkerStyle.FillColor = Colors.Yellow;
+            mp1.MarkerStyle.LineColor = Colors.Blue;
+            mp1.MarkerStyle.FillColor = Colors.Green.WithAlpha(.2);
+            mp1.MarkerStyle.OutlineColor = Colors.Green;
+            mp1.MarkerStyle.OutlineWidth = 2;
 
             var mp2 = plt.Add.Marker(x: .5, y: -i);
             mp2.MarkerStyle.Shape = markerShapes[i];
             mp2.MarkerStyle.Size = 20;
             mp2.MarkerStyle.LineWidth = 2;
+            mp2.MarkerStyle.OutlineWidth = 2;
+            mp2.MarkerStyle.OutlineColor = palette.GetColor(i);
             mp2.MarkerStyle.LineColor = palette.GetColor(i);
             mp2.MarkerStyle.FillColor = palette.GetColor(i).WithAlpha(.3);
-            mp2.MarkerStyle.FillOutline = true;
 
             var txt = plt.Add.Text(markerShapes[i].ToString(), 1, -i);
             txt.LabelAlignment = Alignment.MiddleLeft;