+++ CHANGELOG.md
@@ -7,6 +7,7 @@ _Not yet on NuGet..._
 * Axis Line: Set default line width to 1 which also improved default appearance of crosshair (#3752) @fdesordi
 * Rendering: Copy the plottable list inside the render loop to facilitate plottable list modification mid-render (#3753) @ZSYMAX
 * Controls: Exposed `ZoomRectangle.LineStyle` setter to support advanced customization of middle-click-drag zoom rectangle (#3754) @Graat
+* Markers: Separate `LineColor`, `LineWidth`, etc. from `OutlineColor`, `OutlineWidth`, etc. to allow separate customization of line-based vs. fill-based marker shapes (#3755, #3716) @CD-SailingPerf
 
 ## ScottPlot 5.0.31
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-04-27_